requirements.txt
# Production dependencies
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
alembic==1.15.2
pydantic-settings==2.8.1
azure-monitor-opentelemetry==1.6.7
